requirements.txt
# File parsing libraries
pdfplumber==0.10.3
python-docx==1.1.0

# ML/LLM libraries
google-generativeai==0.3.2

# Environment management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
typing-extensions==4.9.0